Ensuring that clients can be associated with a single user
Updating the factory too to ensure this happens on creation

Resolved:
(FEATURE) Currently, any logged-in user can view all of the system's clients, including those created by other users. Users are obviously not happy with that. Can you make it so that a single Client only belongs to one User?